feat: add pypi distrubution
