+++ .github/workflows/ci.yml
@@ -10,6 +10,10 @@ on:
     paths-ignore:
       - "**.md"
 
+concurrency:
+  group: ci-${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 env:
   RUSTFLAGS: "-D warnings"
 
@@ -67,6 +71,7 @@ jobs:
           --all-features
           --locked
           --workspace
+          --exclude beancount-lsp-python
           --lcov --output-path lcov.info
           -- --include-ignored"
         env:
@@ -104,6 +109,12 @@ jobs:
       - name: Setup | Checkout
         uses: actions/checkout@v6
 
+      - name: Setup | Python (ubuntu)
+        if: matrix.os == 'ubuntu-latest'
+        uses: actions/setup-python@v6
+        with:
+          python-version: "3.11"
+
 
       # Install all the required dependencies for testing
       - name: Setup | Rust
@@ -126,6 +137,7 @@ jobs:
           --all-features
           --locked
           --workspace
+          --exclude beancount-lsp-python
           --lcov --output-path lcov.info
           -- --include-ignored"
         env:
@@ -152,6 +164,50 @@ jobs:
           toolchain: nightly
       - run: cargo doc --all-features --no-deps --workspace
 
+  python-wrapper-build:
+    uses: ./.github/workflows/python-wrapper-build.yml
+
+  python-wheel-test:
+    name: Python Wheel Test
+    runs-on: ${{ matrix.runner }}
+    needs:
+      - python-wrapper-build
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          - runner: windows-latest
+          - runner: macos-15-intel
+          - runner: macos-15
+          - runner: ubuntu-24.04
+          - runner: ubuntu-24.04-arm
+    steps:
+      - name: Setup | Checkout
+        uses: actions/checkout@v6
+
+      - name: Setup | Python
+        uses: actions/setup-python@v6
+        with:
+          python-version: "3.11"
+
+      - name: Setup | uv
+        uses: astral-sh/setup-uv@v7
+        with:
+          activate-environment: "true"
+
+      - name: Download | Python wheels
+        uses: actions/download-artifact@v4
+        with:
+          pattern: python-wheels-*
+          path: dist
+          merge-multiple: true
+
+      - name: Install | Python extension
+        run: uv pip install --system --find-links dist beancount-language-server
+
+      - name: Run | CLI help
+        run: beancount-language-server --help
+
   # Consolidated job for required checks (used for branch protection)
   required-checks:
     name: Required Checks
@@ -160,6 +216,8 @@ jobs:
     needs:
       - rustfmt
       - clippy
+      - python-wrapper-build
+      - python-wheel-test
       - test-ubuntu-stable
     steps:
       - name: Check if all required jobs succeeded
@@ -170,6 +228,14 @@ jobs:
           fi
           if [[ "${{ needs.clippy.result }}" != "success" ]]; then
             echo "clippy failed"
+            exit 1
+          fi
+          if [[ "${{ needs.python-wrapper-build.result }}" != "success" ]]; then
+            echo "python-wrapper-build failed"
+            exit 1
+          fi
+          if [[ "${{ needs.python-wheel-test.result }}" != "success" ]]; then
+            echo "python-wheel-test failed"
             exit 1
           fi
           if [[ "${{ needs.test-ubuntu-stable.result }}" != "success" ]]; then
