+++ .github/workflows/codeql-analysis.yml
@@ -53,7 +53,12 @@ jobs:
         # Details on CodeQL's query packs refer to : https://docs.github.com/en/code-security/code-scanning/automatically-scanning-your-code-for-vulnerabilities-and-errors/configuring-code-scanning#using-queries-in-ql-packs
         # queries: security-extended,security-and-quality
 
-        
+    - name: Setup | Python
+      if: matrix.language == 'cpp'
+      uses: actions/setup-python@v6
+      with:
+        python-version: "3.x"
+
     # Build step - custom for Rust, autobuild for others
     - name: Build Rust
       if: matrix.language == 'cpp'
