+++ .gitignore
@@ -8,6 +8,7 @@ lcov.info
 
 result
 *.py[cod]
+*.so
 .serena
 beancount-language-server.log
 .envrc
