+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     "crates/lsp",
+    "crates/pylsp",
 ]
 resolver = "2"
