+++ python/beancount_language_server/__init__.py
@@ -1,0 +1,3 @@
+from .wrapper import main
+
+__all__ = ["main"]
