+++ python/beancount_language_server/_beancount_lsp.pyi
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from typing import Sequence
+
+
+def main(args: Sequence[str]) -> int: ...
